Create Exercise 2 b: Square Grid
Exercise 2 b: Square Grid ☑️
With the pack layout manager, Create the following labels inside the frames. A and B inside the left frame. C and D inside the right frame
Using Pack() align A and C at the top and B and D at the bottom of the frames that contain them
Support resizing. Use the ‘expand’ and ‘fill’ attributes of the pack method to make the labels grow and expand into the available space.
Assign border value as 5 to the frames